move listEntities
